Core/AreaTriggers: Fixed areatrigger spline duration calculation
Port From (https://github.com/TrinityCore/TrinityCore/commit/5e8adaabcc1bef7563036d3765a0c13894e860f7)

+++ Source/Game/Networking/Packets/MovementPackets.cs
@@ -342,7 +342,7 @@ namespace Game.Networking.Packets
             }
         }
 
-        public static void WriteCreateObjectAreaTriggerSpline(Spline<int> spline, WorldPacket data)
+        public static void WriteCreateObjectAreaTriggerSpline(Spline<float> spline, WorldPacket data)
         {
             data.WriteBits(spline.GetPoints().Length, 16);
             foreach (var point in spline.GetPoints())
